SK-818: Refactor UnlockVestedFunds and show fee
