+++ src/components/Balance/Balance.test.tsx
@@ -1,11 +1,11 @@
 import userEvent from '@testing-library/user-event';
 
 import { render, screen } from '../../testing/testing';
-import { hasVestedFunds } from '../../utilities/vesting/vesting';
+import { hasVestedFunds } from '../../utilities/hasVestedFunds/hasVestedFunds';
 
 import { Balance } from './Balance';
 
-jest.mock('../../utilities/vesting/vesting');
+jest.mock('../../utilities/hasVestedFunds/hasVestedFunds');
 jest.mocked(hasVestedFunds).mockResolvedValue(false);
 
 const mockAddress = '4tDjyLy2gESkLzvaLnpbn7N61VgnwAhqnTHsPPFAwaZjGwP1';
