+++ src/components/BalanceUpdateLink/BalanceUpdateLink.test.tsx
@@ -1,10 +1,10 @@
 import { act, render } from '../../testing/testing';
 
-import { hasVestedFunds } from '../../utilities/vesting/vesting';
+import { hasVestedFunds } from '../../utilities/hasVestedFunds/hasVestedFunds';
 
 import { BalanceUpdateLink } from './BalanceUpdateLink';
 
-jest.mock('../../utilities/vesting/vesting');
+jest.mock('../../utilities/hasVestedFunds/hasVestedFunds');
 jest.mocked(hasVestedFunds).mockResolvedValue(false);
 
 const mockAddress = '4tDjyLy2gESkLzvaLnpbn7N61VgnwAhqnTHsPPFAwaZjGwP1';
