+++ src/views/UnlockVestedFunds/UnlockVestedFunds.module.css
@@ -11,14 +11,13 @@
   margin: 0 0 0.8125rem;
 }
 
-.avatar {
-  composes: avatar from '../../components/Avatar/Avatar.module.css';
-  align-self: center;
-}
-
-.name {
-  composes: identityText from '../../components/Typography/Typography.module.css';
-  margin: 0.5rem 0 0.25rem;
+.fee {
+  composes: smallText from '../../components/Typography/Typography.module.css';
+  font-size: 1.25rem;
+  margin: 0.5rem 0 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
 .explanation {
