+++ src/views/UnlockVestedFunds/UnlockVestedFunds.module.css.d.ts
@@ -2,8 +2,7 @@ export const __esModule: true;
 export const container: string;
 export const heading: string;
 export const subline: string;
-export const avatar: string;
-export const name: string;
+export const fee: string;
 export const explanation: string;
 export const errorTooltip: string;
 export const buttonsLine: string;
