+++ src/views/UnlockVestedFunds/UnlockVestedFunds.test.tsx
@@ -1,7 +1,14 @@
+import { BalanceUtils } from '@kiltprotocol/core';
+
 import { identitiesMock, render } from '../../testing/testing';
 import { waitForGetPassword } from '../../channels/SavedPasswordsChannels/SavedPasswordsChannels.mock';
 
+import { useAsyncValue } from '../../utilities/useAsyncValue/useAsyncValue';
+
 import { UnlockVestedFunds } from './UnlockVestedFunds';
+
+jest.mock('../../utilities/useAsyncValue/useAsyncValue');
+jest.mocked(useAsyncValue).mockReturnValue(BalanceUtils.toFemtoKilt(0.01));
 
 describe('UnlockVestedFunds', () => {
   it('should render', async () => {
