+++ src/views/UnlockVestedFunds/UnlockVestedFunds.tsx
@@ -1,27 +1,43 @@
-import { useState, useCallback } from 'react';
+import { useCallback } from 'react';
 import { browser } from 'webextension-polyfill-ts';
 import { Link } from 'react-router-dom';
+
+import { BlockchainApiConnection } from '@kiltprotocol/chain-helpers';
+
+import BN from 'bn.js';
 
 import * as styles from './UnlockVestedFunds.module.css';
 
 import { Identity } from '../../utilities/identities/types';
 import { paths } from '../paths';
 
-import {
-  insufficientFunds,
-  signVest,
-  submitVest,
-} from '../../utilities/vesting/vesting';
-
-import { Avatar } from '../../components/Avatar/Avatar';
 import { TxStatusModal } from '../../components/TxStatusModal/TxStatusModal';
 import { LinkBack } from '../../components/LinkBack/LinkBack';
 import { Stats } from '../../components/Stats/Stats';
 import {
-  PasswordError,
   PasswordField,
   usePasswordField,
 } from '../../components/PasswordField/PasswordField';
+import { IdentitySlide } from '../../components/IdentitySlide/IdentitySlide';
+import { KiltAmount } from '../../components/KiltAmount/KiltAmount';
+import { useSubmitStates } from '../../utilities/useSubmitStates/useSubmitStates';
+import { useAsyncValue } from '../../utilities/useAsyncValue/useAsyncValue';
+import { getKeypairBySeed } from '../../utilities/identities/identities';
+
+async function getUnsignedExtrinsic() {
+  const blockchain = await BlockchainApiConnection.getConnectionOrConnect();
+  const { api } = blockchain;
+
+  return api.tx.vesting.vest();
+}
+
+export async function getFee(): Promise<BN> {
+  const fakeSeed = new Uint8Array(32);
+  const keypair = getKeypairBySeed(fakeSeed);
+
+  const unsigned = await getUnsignedExtrinsic();
+  return (await unsigned.paymentInfo(keypair)).partialFee;
+}
 
 interface Props {
   identity: Identity;
@@ -30,48 +46,24 @@ interface Props {
 export function UnlockVestedFunds({ identity }: Props): JSX.Element {
   const t = browser.i18n.getMessage;
 
-  const [txStatus, setTxStatus] = useState<
-    'pending' | 'success' | 'error' | null
-  >(null);
-
-  const [txHash, setTxHash] = useState<string>();
-
-  const [error, setError] = useState<string | null>(null);
+  const fee = useAsyncValue(getFee, []);
 
   const passwordField = usePasswordField();
+
+  const { submit, modalProps, submitting, unpaidCosts } = useSubmitStates();
 
   const handleSubmit = useCallback(
     async (event) => {
       event.preventDefault();
 
-      try {
-        const { keypair } = await passwordField.get(event);
+      const { keypair } = await passwordField.get(event);
 
-        setTxStatus('pending');
+      const unsigned = await getUnsignedExtrinsic();
 
-        const hash = await signVest(keypair);
-        setTxHash(hash);
-
-        await submitVest(hash);
-
-        setTxStatus('success');
-      } catch (error) {
-        if (error instanceof PasswordError) {
-          return;
-        }
-        setTxStatus('error');
-        console.error(error);
-        if (error instanceof Error && error.message === insufficientFunds) {
-          setError(t('view_UnlockVestedFunds_insufficient_funds'));
-        }
-      }
+      await submit(keypair, unsigned);
     },
-    [t, passwordField],
+    [passwordField, submit],
   );
-
-  const closeModal = useCallback(() => {
-    setTxStatus(null);
-  }, []);
 
   return (
     <form
@@ -82,8 +74,14 @@ export function UnlockVestedFunds({ identity }: Props): JSX.Element {
       <h1 className={styles.heading}>{t('view_UnlockVestedFunds_heading')}</h1>
       <p className={styles.subline}>{t('view_UnlockVestedFunds_subline')}</p>
 
-      <Avatar identity={identity} className={styles.avatar} />
-      <p className={styles.name}>{identity.name}</p>
+      <IdentitySlide identity={identity} />
+
+      {fee && (
+        <p className={styles.fee}>
+          {t('view_DidEndpointsSign_fee')}
+          <KiltAmount amount={fee} type="costs" smallDecimals />
+        </p>
+      )}
 
       <p className={styles.explanation}>
         {t('view_UnlockVestedFunds_explanation')}
@@ -95,25 +93,18 @@ export function UnlockVestedFunds({ identity }: Props): JSX.Element {
         <Link to={paths.home} className={styles.cancel}>
           {t('common_action_cancel')}
         </Link>
-        <button type="submit" className={styles.submit}>
+        <button type="submit" className={styles.submit} disabled={submitting}>
           {t('view_UnlockVestedFunds_CTA')}
         </button>
         <output
           className={styles.errorTooltip}
-          hidden={!error || Boolean(txStatus)}
+          hidden={!unpaidCosts || Boolean(modalProps)}
         >
-          {error}
+          {t('view_UnlockVestedFunds_insufficient_funds', [unpaidCosts])}
         </output>
       </p>
 
-      {txStatus && (
-        <TxStatusModal
-          identity={identity}
-          status={txStatus}
-          txHash={txHash}
-          onDismissError={closeModal}
-        />
-      )}
+      {modalProps && <TxStatusModal {...modalProps} identity={identity} />}
 
       <LinkBack />
       <Stats />
